requirements.txt
appdirs==1.4.3
Jinja2==2.9.5
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
PyYAML==3.12
six==1.10.0
